requirements.txt
urllib3
rich
colorama
requests
bs4
jinja2
dnspython
aiohttp
whois
pyOpenSSL
